Make frontpage faster
