+++ apps/web/src/lib/is-bedpres.ts
@@ -1,0 +1,7 @@
+import { type fetchHomeHappenings } from "@/sanity/happening/requests";
+
+export const isBedpres = (
+  happening: Awaited<ReturnType<typeof fetchHomeHappenings>>[number],
+): happening is Awaited<ReturnType<typeof fetchHomeHappenings<"bedpres">>>[number] => {
+  return happening.happeningType === "bedpres";
+};
